requirements.txt
rzpipe
langchain
langchain-openai
langchain-deepseek
pyyaml
httpx
pydantic
python-dotenv
fastapi
uvicorn
python-multipartsqlalchemy
aiofiles